+++ assets/src/components/layout/Header.tsx
@@ -1,6 +1,5 @@
 import { Flex } from '@pluralsh/design-system'
-import { useNavigate } from 'react-router-dom'
-import styled, { useTheme } from 'styled-components'
+import styled from 'styled-components'
 
 import BillingLegacyUserMessage from 'components/billing/BillingLegacyUserMessage'
 
@@ -10,9 +9,6 @@ import { ChatbotLauncher } from 'components/ai/chatbot/Chatbot'
 import DemoBanner from './DemoBanner'
 import { HeaderProjectSelect } from './HeaderProjectSelect'
 import { ProfileMenu } from './ProfileMenu'
-
-const APP_ICON_LIGHT = '/plural-logo.png'
-const APP_ICON_DARK = '/plural-logo-white.png'
 
 const HeaderSC = styled.div(({ theme }) => ({
   backgroundColor: theme.colors['fill-accent'],
@@ -26,27 +22,11 @@ const HeaderContentSC = styled.div(({ theme }) => ({
   padding: `${theme.spacing.xsmall}px ${theme.spacing.large}px`,
 }))
 
-const LogoSC = styled.img(({ theme }) => ({
-  cursor: 'pointer',
-  marginLeft: -7 /* Optically center with sidebar buttons */,
-  height: 32,
-  marginTop: theme.spacing.xxsmall,
-  marginBottom: theme.spacing.xxsmall,
-}))
-
 export default function Header() {
-  const navigate = useNavigate()
-  const theme = useTheme()
-
   return (
     <HeaderSC>
       <DemoBanner />
       <HeaderContentSC>
-        <LogoSC
-          src={theme.mode === 'light' ? APP_ICON_LIGHT : APP_ICON_DARK}
-          alt="Plural console"
-          onClick={() => navigate('/')}
-        />
         <HeaderProjectSelect />
         <Flex grow={1} />
         <BillingLegacyUserMessage />
